# Assignment 12/docs/openapi.yaml
openapi: 3.0.0
info:
  title: AI-Powered Smart Appointment Booking System API
  description: |
    REST API for managing patients, doctors, and appointments in an AI-Powered Smart Appointment Booking System.
    
    This API provides endpoints for creating, reading, updating, and deleting patients, doctors, and appointments.
    It also includes specialized endpoints for business workflows such as appointment cancellation.
    
    The API enforces several business rules:
    - Patients can't book more than 3 appointments in a single day
    - Appointments must be booked at least 24 hours in advance
    - Cancellations must be made at least 6 hours before the appointment
    - Doctors cannot be double-booked for appointments
  version: 1.0.0
  contact:
    name: Aphiwe Andrew Ncama
    email: ncayiyane@gmail.com
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
  /api/patients:
    get:
      summary: Get all patients
      description: Retrieves a list of all patients registered in the system.
      tags:
        - Patients
      responses:
        '200':
          description: List of patients successfully retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'
              example:
                - id: "patient-123"
                  name: "John Doe"
                  contact_info:
                    email: "john.doe@example.com"
                    phone: "123-456-7890"
                    address: "123 Main St, Anytown, USA"
                - id: "patient-456"
                  name: "Jane Smith"
                  contact_info:
                    email: "jane.smith@example.com"
                    phone: "987-654-3210"
                    address: "456 Oak St, Anytown, USA"
    post:
      summary: Create a new patient
      description: Creates a new patient in the system. Each patient must have a unique email address.
      tags:
        - Patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
            example:
              name: "John Doe"
              contact_info:
                email: "john.doe@example.com"
                phone: "123-456-7890"
                address: "123 Main St, Anytown, USA"
      responses:
        '201':
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
              example:
                id: "patient-123"
                name: "John Doe"
                contact_info:
                  email: "john.doe@example.com"
                  phone: "123-456-7890"
                  address: "123 Main St, Anytown, USA"
        '400':
          description: Bad request - invalid input or duplicate email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                detail: "Patient with email john.doe@example.com already exists"
  /api/patients/{patient_id}:
    get:
      summary: Get a patient by ID
      tags:
        - Patients
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a patient
      tags:
        - Patients
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a patient
      tags:
        - Patients
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Patient deleted successfully
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/patients/search/{name}:
    get:
      summary: Search for patients by name
      tags:
        - Patients
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of patients matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientResponse'
  /api/doctors:
    get:
      summary: Get all doctors
      tags:
        - Doctors
      responses:
        '200':
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorResponse'
    post:
      summary: Create a new doctor
      tags:
        - Doctors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorCreate'
      responses:
        '201':
          description: Doctor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/doctors/{doctor_id}:
    get:
      summary: Get a doctor by ID
      tags:
        - Doctors
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Doctor found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update a doctor
      tags:
        - Doctors
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorUpdate'
      responses:
        '200':
          description: Doctor updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a doctor
      tags:
        - Doctors
      parameters:
        - name: doctor_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Doctor deleted successfully
        '404':
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/doctors/specialization/{specialization}:
    get:
      summary: Search for doctors by specialization
      tags:
        - Doctors
      parameters:
        - name: specialization
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of doctors matching the specialization
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorResponse'
  /api/doctors/search/{name}:
    get:
      summary: Search for doctors by name
      tags:
        - Doctors
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of doctors matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorResponse'
...
components:
  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail
    ContactInfo:
      type: object
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
      required:
        - email
        - phone
        - address
    PatientCreate:
      type: object
      properties:
        name:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
      required:
        - name
        - contact_info
    PatientResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
      required:
        - id
        - name
        - contact_info
    PatientUpdate:
      type: object
      properties:
        name:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
    DoctorCreate:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
      required:
        - name
        - specialization
        - contact_info
    DoctorResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        specialization:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
      required:
        - id
        - name
        - specialization
        - contact_info
    DoctorUpdate:
      type: object
      properties:
        name:
          type: string
        specialization:
          type: string
        contact_info:
          $ref: '#/components/schemas/ContactInfo'
...